+++ persistence/src/main/java/de/starwit/persistence/observatory/entity/JobType.java
@@ -2,5 +2,6 @@ package de.starwit.persistence.observatory.entity;
 
 public enum JobType {
     LINE_CROSSING,
-    AREA_OCCUPANCY;
+    AREA_OCCUPANCY,
+    FLOW;
 }