+++ service/src/main/java/de/starwit/service/analytics/MetadataService.java
@@ -58,5 +58,15 @@ public class MetadataService implements ServiceInterface<MetadataEntity, Metadat
     public void delete(Long id) {
         throw new UnsupportedOperationException("Delete is not supported");
     }
+
+    public MetadataEntity findFirstByName(String name) {
+        MetadataEntity metadata = metadataRepository.findFirstByName(name);
+        return metadata;
+    }
+
+    public MetadataEntity findFirstByNameAndClassification(String name, String classification) {
+        MetadataEntity metadata = metadataRepository.findFirstByNameAndClassification(name, classification);
+        return metadata;
+    }
     
 }