+++ service/src/main/java/de/starwit/service/geojson/GeoJsonService.java
@@ -6,9 +6,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
-import de.starwit.service.jobs.AreaOccupancyObservation;
-import de.starwit.service.jobs.LineCrossingObservation;
+import de.starwit.service.jobs.areaoccupancy.AreaOccupancyObservation;
+import de.starwit.service.jobs.linecrossing.LineCrossingObservation;
 
+/**
+ * Service for vehicle positions in GeoJSON format in order to demonstrate real-time visualization e.g. for ArcGIS.
+ */
 @Service
 public class GeoJsonService {