+++ service/src/main/java/de/starwit/service/jobs/linecrossing/LineCrossingObservation.java
@@ -1,4 +1,4 @@
-package de.starwit.service.jobs;
+package de.starwit.service.jobs.linecrossing;
 
 import de.starwit.persistence.analytics.entity.Direction;
 import de.starwit.persistence.observatory.entity.ObservationJobEntity;