+++ service/src/main/java/de/starwit/service/jobs/linecrossing/LineCrossingRunner.java
@@ -1,4 +1,4 @@
-package de.starwit.service.jobs;
+package de.starwit.service.jobs.linecrossing;
 
 import java.time.Duration;
 import java.util.ArrayList;
@@ -17,17 +17,21 @@ import org.springframework.data.redis.stream.StreamMessageListenerContainer;
 import org.springframework.data.redis.stream.Subscription;
 import org.springframework.stereotype.Service;
 
+import de.starwit.persistence.observatory.entity.JobType;
 import de.starwit.persistence.observatory.entity.ObservationJobEntity;
 import de.starwit.service.analytics.LineCrossingService;
 import de.starwit.service.geojson.GeoJsonService;
+import de.starwit.service.jobs.JobInterface;
+import de.starwit.service.jobs.RunnerInterface;
 import de.starwit.service.observatory.ObservationJobService;
-import de.starwit.service.sae.SaeDetectionDto;
 import de.starwit.service.sae.SaeMessageListener;
 import jakarta.annotation.PostConstruct;
 
 @Service
-public class LineCrossingRunner {
+public class LineCrossingRunner implements RunnerInterface {
     private final Logger log = LoggerFactory.getLogger(this.getClass());
+
+    public static final JobType JOB_TYPE = JobType.LINE_CROSSING;
 
     @Value("${lineCrossing.targetWindowSize:1s}")
     private Duration TARGET_WINDOW_SIZE;    
@@ -62,8 +66,14 @@ public class LineCrossingRunner {
         refreshJobs();
     }
 
+    @Override
+    public List<? extends JobInterface> getActiveJobs() {
+        return this.activeJobs;
+    }
+
+    @Override
     public void refreshJobs() {
-        log.debug("Refreshing jobs");
+        log.info("Refreshing jobs");
 
         for (Subscription activeSub : activeSubscriptions) {
             streamListenerContainer.remove(activeSub);
@@ -71,7 +81,7 @@ public class LineCrossingRunner {
         
         this.activeSubscriptions = new ArrayList<>();
 
-        List<ObservationJobEntity> enabledJobEntites = observationJobService.findActiveLineCrossingJobs();
+        List<ObservationJobEntity> enabledJobEntites = observationJobService.findActiveJobs(JOB_TYPE);
         log.info("Enabled jobs: " + enabledJobEntites.stream().map(j -> j.getName()).collect(Collectors.joining(",")));
 
         this.activeJobs = enabledJobEntites.stream().map(jobEntity -> new LineCrossingJob(jobEntity, TARGET_WINDOW_SIZE, this::storeObservation)).toList();
@@ -82,14 +92,6 @@ public class LineCrossingRunner {
             StreamOffset<String> streamOffset = StreamOffset.create(streamKey, ReadOffset.lastConsumed());
             Subscription redisSubscription = streamListenerContainer.receive(streamOffset, saeMessageListener);
             activeSubscriptions.add(redisSubscription);
-        }
-    }
-
-    public void messageHandler(SaeDetectionDto dto) {
-        for (LineCrossingJob lineJob : activeJobs) {
-            if (lineJob.getConfigEntity().getCameraId().equals(dto.getCameraId())) {
-                lineJob.processNewDetection(dto);
-            }
         }
     }