+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 Changelog of observatory.
 
+## Current Version ()
+
+### Features
+
+-  ab#1981: added rest service to set occupancy count ([0ab24](https://github.com/starwit/observatory/commit/0ab24137b57e5f0) Anett Hübner)  
+-  ab#1981: added flow as job type ([ed415](https://github.com/starwit/observatory/commit/ed415b06472ab8a) Anett Hübner)  
+
 ## refs/tags/1.3.1 (2025-01-23)
 
 ### Bug Fixes